6860: Updated approved file and moved S3BucketName
