+++ java/deployment/cdk/src/main/java/sleeper/cdk/stack/AthenaStack.java
@@ -36,11 +36,11 @@ import software.constructs.Construct;
 import sleeper.cdk.artefacts.SleeperInstanceArtefacts;
 import sleeper.cdk.lambda.SleeperLambdaCode;
 import sleeper.cdk.stack.core.LoggingStack.LogGroupRef;
+import sleeper.cdk.util.S3BucketName;
 import sleeper.cdk.util.Utils;
 import sleeper.core.deploy.LambdaHandler;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.util.EnvironmentUtils;
-import sleeper.core.util.S3BucketName;
 
 import java.util.List;
 import java.util.Map;
