+++ java/deployment/cdk/src/main/java/sleeper/cdk/stack/bulkexport/BulkExportStack.java
@@ -39,12 +39,12 @@ import sleeper.cdk.artefacts.containers.SleeperEcsImages;
 import sleeper.cdk.lambda.SleeperLambdaCode;
 import sleeper.cdk.stack.SleeperCoreStacks;
 import sleeper.cdk.stack.core.LoggingStack.LogGroupRef;
+import sleeper.cdk.util.S3BucketName;
 import sleeper.cdk.util.Utils;
 import sleeper.core.deploy.LambdaHandler;
 import sleeper.core.properties.instance.CdkDefinedInstanceProperty;
 import sleeper.core.properties.instance.InstanceProperties;
 import sleeper.core.util.EnvironmentUtils;
-import sleeper.core.util.S3BucketName;
 
 import java.util.Arrays;
 import java.util.Collections;
