+++ java/deployment/cdk/src/main/java/sleeper/cdk/util/NewInstanceValidator.java
@@ -22,7 +22,6 @@ import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.NoSuchBucketException;
 
 import sleeper.core.properties.instance.InstanceProperties;
-import sleeper.core.util.S3BucketName;
 
 import static sleeper.core.properties.instance.CdkDefinedInstanceProperty.ACCOUNT;
 import static sleeper.core.properties.instance.CommonProperty.ID;
