+++ java/deployment/cdk/src/main/java/sleeper/cdk/util/S3BucketName.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.core.util;
+package sleeper.cdk.util;
 
 import java.util.Locale;
