+++ java/deployment/cdk/src/test/java/sleeper/cdk/default-instance.approved.json
@@ -7685,7 +7685,7 @@
               "BlockPublicAcls": true,
               "IgnorePublicAcls": true
             },
-            "BucketName": "sleeper-test-instance-table-data",
+            "BucketName": "sleeper-test-instance-table-data-test-account",
             "BucketEncryption": {
               "ServerSideEncryptionConfiguration": [
                 {
@@ -8294,7 +8294,7 @@
               "BlockPublicAcls": true,
               "IgnorePublicAcls": true
             },
-            "BucketName": "sleeper-test-instance-config",
+            "BucketName": "sleeper-test-instance-config-test-account",
             "BucketEncryption": {
               "ServerSideEncryptionConfiguration": [
                 {
@@ -13175,7 +13175,7 @@
               "BlockPublicAcls": true,
               "IgnorePublicAcls": true
             },
-            "BucketName": "sleeper-test-instance-bulk-import",
+            "BucketName": "sleeper-test-instance-bulk-import-test-account",
             "BucketEncryption": {
               "ServerSideEncryptionConfiguration": [
                 {
@@ -22965,7 +22965,7 @@
               "BlockPublicAcls": true,
               "IgnorePublicAcls": true
             },
-            "BucketName": "sleeper-test-instance-query-results",
+            "BucketName": "sleeper-test-instance-query-results-test-account",
             "BucketEncryption": {
               "ServerSideEncryptionConfiguration": [
                 {
